requirements.txt
setuptools==65.5.0
opencv-python==4.5.4.60
stable_baselines3==1.7.0
gym
wandb
hydra-core
colorama
numpy
pytest
scipy
pytorch3d
torchgeometry
moviepy
imageio
tensorboard
